requirements.txt
openai
faiss-cpu
tiktoken
pytesseract
PyMuPDF
pillow
fastapi
uvicorn
scikit-learn
nest_asyncio
streamlit
python-dotenv
numpy